feat/#304: 전역 프로바이더 사용을 위한 콘텐츠 분리
